Make video loading more efficient and bug-free
Fixes https://github.com/sakuga-video/guesser/issues/32
Side effect: it may show videos multiple times in a row accidentally, because the random roll doesn't take into account previous random rolls. That's a small price to pay for the code not getting caught in rapid-reuqest loops when it goes beyond the limits of the number of videos, or finds a bunch of jpgs in a row.